Handle both end of track and playback
Depending on if we're running as a DSP with reset/no-reset or regular
playback in conjunction with other DSPs in the chain, we may have
different sequences of `on_chunk`, `on_endoftrack`, `on_endofplayback`.

This commit takes the same action of inserting post-track silence in
both end-of-track and end-of-playback, but only if there has been an
intervening chunk.

This should fix issue #5.